Issue 17 and 18
split stores into sub stores
added read only view .
