+++ frontend/src/components/Data/Data.js
@@ -1,24 +1,22 @@
 import React, {Component} from "react";
-
 import InputFormList from "./InputFormList";
 import InputButtons from "./InputButtons";
 import "./Data.css";
-
 import {inject, observer} from "mobx-react";
 import DatasetScatterplot from "./DatasetScatterplot";
 
-@inject("store")
+@inject("dataStore")
 @observer
 class Data extends Component {
     render() {
-        const {store} = this.props;
+        const {dataStore} = this.props;
         return (
             <div className="container position-static">
                 <div className="row">
                     <InputButtons />
-                    {store.getDataLength ? <InputFormList /> : null}
+                    {dataStore.getDataLength ? <InputFormList /> : null}
 
-                    {store.getDataLength ? <DatasetScatterplot /> : null}
+                    {dataStore.getDataLength ? <DatasetScatterplot /> : null}
                 </div>
             </div>
         );
