+++ frontend/src/components/Main/ModelsCheckBoxList.js
@@ -2,25 +2,28 @@ import React, {Component} from "react";
 
 import {inject, observer} from "mobx-react";
 import ModelsCheckBox from "./ModelsCheckBox";
+import ModelsReadOnly from "./ModelsReadOnly";
 import {toJS} from "mobx";
 
-@inject("store")
+@inject("mainStore")
 @observer
 class ModelsCheckBoxList extends Component {
-    handleCheckbox = e => {
-        let model_name = e.target.name;
-        let checked = e.target.checked;
-        let value = e.target.value;
-        this.props.store.toggleModelsCheckBox(model_name, checked, value);
-    };
     render() {
-        let models = toJS(this.props.store.getModelTypeList());
+        const {mainStore} = this.props,
+            handleCheckbox = e => {
+                let model_name = e.target.name;
+                let checked = e.target.checked;
+                let value = e.target.value;
+                mainStore.toggleModelsCheckBox(model_name, checked, value);
+            },
+            models = toJS(mainStore.getModelTypeList()),
+            isEditSettings = mainStore.getEditSettings();
         return (
             <div>
                 <div className="checkbox-table">
                     <table className="table table-bordered hover">
                         <thead>
-                            {this.props.store.modelsCheckBoxHeaders.map((item, index) => {
+                            {mainStore.modelsCheckBoxHeaders.map((item, index) => {
                                 return [
                                     <tr key={index}>
                                         <th>{item.model}</th>
@@ -28,14 +31,14 @@ class ModelsCheckBoxList extends Component {
                                             return [
                                                 <th key={index} colSpan={dev.colspan}>
                                                     {dev.name}{" "}
-                                                    {dev.name === "Enable" ? (
+                                                    {(dev.name === "Enable") & isEditSettings ? (
                                                         <input
                                                             type="checkbox"
                                                             name={dev.model_name + "-All"}
                                                             onChange={this.handleCheckbox}
                                                         />
                                                     ) : null}
-                                                    &nbsp; &nbsp;
+                                                    &emsp;
                                                     {dev.model_name === "bayesian_model_average"
                                                         ? dev.prior_weight
                                                         : null}
@@ -46,7 +49,11 @@ class ModelsCheckBoxList extends Component {
                                 ];
                             })}
                         </thead>
-                        <ModelsCheckBox models={models} onChange={this.handleCheckbox} />
+                        {isEditSettings ? (
+                            <ModelsCheckBox models={models} onChange={handleCheckbox} />
+                        ) : (
+                            <ModelsReadOnly models={models} />
+                        )}
                     </table>
                 </div>
             </div>
