+++ frontend/src/components/Main/OptionsFormList.js
@@ -2,29 +2,36 @@ import React, {Component} from "react";
 import {inject, observer} from "mobx-react";
 import OptionsForm from "./OptionsForm";
 import {toJS} from "mobx";
+import OptionsReadOnly from "./OptionsReadOnly";
 
-@inject("store")
+@inject("mainStore")
 @observer
 class OptionsFormList extends Component {
-    onChange = e => {
-        e.preventDefault();
-        const {name, value, id} = e.target;
-        let parsedValue = "";
-        if (name == "bmr_value" || name == "tail_probability" || name == "confidence_level") {
-            parsedValue = parseFloat(value);
-        } else {
-            parsedValue = value;
-        }
-        this.props.store.saveOptions(name, parsedValue, id);
-    };
-    deleteOption = (e, val) => {
-        e.preventDefault();
-        this.props.store.deleteOptions(val);
-    };
     render() {
-        const {store} = this.props;
-        let dataset_type = store.analysisForm.dataset_type;
-        let options = toJS(store.options);
+        const {mainStore} = this.props,
+            onChange = e => {
+                e.preventDefault();
+                const {name, value, id} = e.target;
+                let parsedValue = "";
+                if (
+                    name == "bmr_value" ||
+                    name == "tail_probability" ||
+                    name == "confidence_level"
+                ) {
+                    parsedValue = parseFloat(value);
+                } else {
+                    parsedValue = value;
+                }
+                mainStore.saveOptions(name, parsedValue, id);
+            },
+            deleteOption = (e, val) => {
+                e.preventDefault();
+                mainStore.deleteOptions(val);
+            },
+            dataset_type = mainStore.analysisForm.dataset_type,
+            labels = mainStore.getOptionsLabels(dataset_type),
+            options = toJS(mainStore.options),
+            isEditSettings = mainStore.getEditSettings();
         return (
             <form>
                 <div className="row" style={{marginTop: 20}}>
@@ -33,41 +40,46 @@ class OptionsFormList extends Component {
                             <table className="table table-bordered">
                                 <thead>
                                     <tr>
-                                        <th>Option Set #</th>
-                                        {dataset_type === "C" ? <th>BMR Type</th> : null}
-                                        {dataset_type === "D" ? <th>Risk Type</th> : null}
-                                        <th>BMRF</th>
-                                        {dataset_type === "C" ? <th>Tail Probability</th> : null}
-                                        <th>Confidence Level</th>
-                                        {dataset_type === "C" ? <th>Distribution</th> : null}
-                                        {dataset_type === "C" ? <th>Variance</th> : null}
-                                        {dataset_type === "C" ? (
-                                            <th>Polynomial Restriction</th>
-                                        ) : null}
-                                        <th>Background</th>
-                                        <th></th>
+                                        {labels.map((item, index) => {
+                                            return [<th key={index}>{item.label}</th>];
+                                        })}
                                     </tr>
                                 </thead>
-                                <tbody>
-                                    {options.map((item, id) => (
-                                        <OptionsForm
-                                            className="form-control"
-                                            key={id}
-                                            item={item}
-                                            dataset_type={dataset_type}
-                                            idx={id}
-                                            onchange={this.onChange}
-                                            delete={this.deleteOption.bind(this)}
-                                        />
-                                    ))}
-                                </tbody>
+                                {isEditSettings ? (
+                                    <tbody>
+                                        {options.map((item, id) => (
+                                            <OptionsForm
+                                                className="form-control"
+                                                key={id}
+                                                item={item}
+                                                dataset_type={dataset_type}
+                                                idx={id}
+                                                onchange={onChange}
+                                                delete={deleteOption.bind(this)}
+                                            />
+                                        ))}
+                                    </tbody>
+                                ) : (
+                                    <tbody>
+                                        {options.map((item, id) => (
+                                            <OptionsReadOnly
+                                                className="form-control"
+                                                key={id}
+                                                idx={id}
+                                                item={item}
+                                            />
+                                        ))}
+                                    </tbody>
+                                )}
                             </table>
-                            <button
-                                type="button"
-                                className="btn btn-primary "
-                                onClick={() => store.createOptions()}>
-                                add option set
-                            </button>
+                            {isEditSettings ? (
+                                <button
+                                    type="button"
+                                    className="btn btn-primary "
+                                    onClick={() => mainStore.createOptions()}>
+                                    Add option set
+                                </button>
+                            ) : null}
                         </div>
                     </div>
                 </div>
